Split core tests into focused modules
- test_config.py: Settings configuration tests
- test_logging.py: Logging setup and output tests
- test_exceptions.py: Custom exception tests
